Add -q option which lets caheui quiet on input operations
If this option is specified, the interpretter does not output
messages to stdout when an input operation is executed.

Background: I have recently added this language to my golf
site (http://golf.shinh.org/). These extra outputs to stdout
makes the problems which require input unsolvable. This patch
would resolve the issue.